Fix constraintError, fix seed-method

+++ PhoneCat/Migrations/201702161335577_sizeAndWeight.cs
@@ -7,10 +7,10 @@ namespace PhoneCat.Migrations
     {
         public override void Up()
         {
-            AddColumn("dbo.Phones", "SizeAndWeight_Height", c => c.Int(nullable: false));
-            AddColumn("dbo.Phones", "SizeAndWeight_Width", c => c.Int(nullable: false));
-            AddColumn("dbo.Phones", "SizeAndWeight_Depth", c => c.Int(nullable: false));
-            AddColumn("dbo.Phones", "SizeAndWeight_Weight", c => c.Int(nullable: false));
+            AddColumn("dbo.Phones", "SizeAndWeight_Height", c => c.Double(nullable: false));
+            AddColumn("dbo.Phones", "SizeAndWeight_Width", c => c.Double(nullable: false));
+            AddColumn("dbo.Phones", "SizeAndWeight_Depth", c => c.Double(nullable: false));
+            AddColumn("dbo.Phones", "SizeAndWeight_Weight", c => c.Double(nullable: false));
         }
         
         public override void Down()
